Remove support for S3 Select
