+++ docs/src/main/sphinx/connector/hive-s3.md
@@ -312,85 +312,3 @@ classpath and must be able to communicate with your custom key management system
 the `org.apache.hadoop.conf.Configurable` interface from the Hadoop Java API, then the Hadoop configuration
 is passed in after the object instance is created, and before it is asked to provision or retrieve any
 encryption keys.
-
-(s3selectpushdown)=
-
-## S3 Select pushdown
-
-S3 Select pushdown enables pushing down projection (SELECT) and predicate (WHERE)
-processing to [S3 Select](https://docs.aws.amazon.com/AmazonS3/latest/API/RESTObjectSELECTContent.html).
-With S3 Select Pushdown, Trino only retrieves the required data from S3 instead
-of entire S3 objects, reducing both latency and network usage.
-
-### Is S3 Select a good fit for my workload?
-
-Performance of S3 Select pushdown depends on the amount of data filtered by the
-query. Filtering a large number of rows should result in better performance. If
-the query doesn't filter any data, then pushdown may not add any additional value
-and the user is charged for S3 Select requests. Thus, we recommend that you
-benchmark your workloads with and without S3 Select to see if using it may be
-suitable for your workload. By default, S3 Select Pushdown is disabled and you
-should enable it in production after proper benchmarking and cost analysis. For
-more information on S3 Select request cost, please see
-[Amazon S3 Cloud Storage Pricing](https://aws.amazon.com/s3/pricing/).
-
-Use the following guidelines to determine if S3 Select is a good fit for your
-workload:
-
-- Your query filters out more than half of the original data set.
-- Your query filter predicates use columns that have a data type supported by
-  Trino and S3 Select.
-  The `TIMESTAMP`, `DECIMAL`, `REAL`, and `DOUBLE` data types are not
-  supported by S3 Select Pushdown. For more information about supported data
-  types for S3 Select, see the
-  [Data Types documentation](https://docs.aws.amazon.com/AmazonS3/latest/dev/s3-glacier-select-sql-reference-data-types.html).
-- Your network connection between Amazon S3 and the Amazon EMR cluster has good
-  transfer speed and available bandwidth. Amazon S3 Select does not compress
-  HTTP responses, so the response size may increase for compressed input files.
-
-### Considerations and limitations
-
-- Only objects stored in JSON format are supported. Objects can be uncompressed,
-  or optionally compressed with gzip or bzip2.
-- The "AllowQuotedRecordDelimiters" property is not supported. If this property
-  is specified, the query fails.
-- Amazon S3 server-side encryption with customer-provided encryption keys
-  (SSE-C) and client-side encryption are not supported.
-- S3 Select Pushdown is not a substitute for using columnar or compressed file
-  formats such as ORC and Parquet.
-
-### Enabling S3 Select pushdown
-
-You can enable S3 Select Pushdown using the `s3_select_pushdown_enabled`
-Hive session property, or using the `hive.s3select-pushdown.enabled`
-configuration property. The session property overrides the config
-property, allowing you enable or disable on a per-query basis. Non-filtering
-queries (`SELECT * FROM table`) are not pushed down to S3 Select,
-as they retrieve the entire object content.
-
-For uncompressed files, S3 Select scans ranges of bytes in parallel. The scan range
-requests run across the byte ranges of the internal Hive splits for the query fragments
-pushed down to S3 Select. Changes in the Hive connector {ref}`performance tuning
-configuration properties <hive-performance-tuning-configuration>` are likely to impact
-S3 Select pushdown performance.
-
-S3 Select can be enabled for TEXTFILE data using the
-`hive.s3select-pushdown.experimental-textfile-pushdown-enabled` configuration property,
-however this has been shown to produce incorrect results. For more information see
-[the GitHub Issue.](https://github.com/trinodb/trino/issues/17775)
-
-### Understanding and tuning the maximum connections
-
-Trino can use its native S3 file system or EMRFS. When using the native FS, the
-maximum connections is configured via the `hive.s3.max-connections`
-configuration property. When using EMRFS, the maximum connections is configured
-via the `fs.s3.maxConnections` Hadoop configuration property.
-
-S3 Select Pushdown bypasses the file systems, when accessing Amazon S3 for
-predicate operations. In this case, the value of
-`hive.s3select-pushdown.max-connections` determines the maximum number of
-client connections allowed for those operations from worker nodes.
-
-If your workload experiences the error *Timeout waiting for connection from
-pool*, increase the value of both `hive.s3select-pushdown.max-connections` and
-the maximum connections configuration for the file system you are using.
