+++ plugin/trino-hive-hadoop2/src/test/java/io/trino/plugin/hive/AbstractTestHiveFileSystemAbfs.java
@@ -66,7 +66,6 @@ public abstract class AbstractTestHiveFileSystemAbfs
                 checkParameter(host, "host"),
                 port,
                 checkParameter(databaseName, "database name"),
-                false,
                 createHdfsConfiguration());
     }
