+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/HivePageSourceProvider.java
@@ -147,7 +147,6 @@ public class HivePageSourceProvider
                 typeManager,
                 hiveSplit.getBucketConversion(),
                 hiveSplit.getBucketValidation(),
-                hiveSplit.isS3SelectPushdownEnabled(),
                 hiveSplit.getAcidInfo(),
                 originalFile,
                 hiveTable.getTransaction(),
@@ -177,7 +176,6 @@ public class HivePageSourceProvider
             TypeManager typeManager,
             Optional<BucketConversion> bucketConversion,
             Optional<BucketValidation> bucketValidation,
-            boolean s3SelectPushdownEnabled,
             Optional<AcidInfo> acidInfo,
             boolean originalFile,
             AcidTransaction transaction,
