+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/HiveSplit.java
@@ -55,7 +55,6 @@ public class HiveSplit
     private final TableToPartitionMapping tableToPartitionMapping;
     private final Optional<BucketConversion> bucketConversion;
     private final Optional<BucketValidation> bucketValidation;
-    private final boolean s3SelectPushdownEnabled;
     private final Optional<AcidInfo> acidInfo;
     private final SplitWeight splitWeight;
 
@@ -76,7 +75,6 @@ public class HiveSplit
             @JsonProperty("tableToPartitionMapping") TableToPartitionMapping tableToPartitionMapping,
             @JsonProperty("bucketConversion") Optional<BucketConversion> bucketConversion,
             @JsonProperty("bucketValidation") Optional<BucketValidation> bucketValidation,
-            @JsonProperty("s3SelectPushdownEnabled") boolean s3SelectPushdownEnabled,
             @JsonProperty("acidInfo") Optional<AcidInfo> acidInfo,
             @JsonProperty("splitWeight") SplitWeight splitWeight)
     {
@@ -110,7 +108,6 @@ public class HiveSplit
         this.tableToPartitionMapping = tableToPartitionMapping;
         this.bucketConversion = bucketConversion;
         this.bucketValidation = bucketValidation;
-        this.s3SelectPushdownEnabled = s3SelectPushdownEnabled;
         this.acidInfo = acidInfo;
         this.splitWeight = requireNonNull(splitWeight, "splitWeight is null");
     }
@@ -213,12 +210,6 @@ public class HiveSplit
     }
 
     @JsonProperty
-    public boolean isS3SelectPushdownEnabled()
-    {
-        return s3SelectPushdownEnabled;
-    }
-
-    @JsonProperty
     public Optional<AcidInfo> getAcidInfo()
     {
         return acidInfo;
@@ -261,7 +252,6 @@ public class HiveSplit
                 .put("forceLocalScheduling", forceLocalScheduling)
                 .put("partitionName", partitionName)
                 .put("deserializerClassName", getDeserializerClassName(schema))
-                .put("s3SelectPushdownEnabled", s3SelectPushdownEnabled)
                 .buildOrThrow();
     }
