+++ plugin/trino-hive/src/main/java/io/trino/plugin/hive/HiveSplitSource.java
@@ -312,7 +312,6 @@ class HiveSplitSource
                         internalSplit.getTableToPartitionMapping(),
                         internalSplit.getBucketConversion(),
                         internalSplit.getBucketValidation(),
-                        internalSplit.isS3SelectPushdownEnabled(),
                         internalSplit.getAcidInfo(),
                         splitWeightProvider.weightForSplitSizeInBytes(splitBytes)));
