+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestBackgroundHiveSplitLoader.java
@@ -821,7 +821,6 @@ public class TestBackgroundHiveSplitLoader
                 Optional.empty(),
                 DataSize.of(512, MEGABYTE),
                 false,
-                false,
                 Optional.empty());
         BackgroundHiveSplitLoader backgroundHiveSplitLoader = backgroundHiveSplitLoader(
                 files,
@@ -863,7 +862,6 @@ public class TestBackgroundHiveSplitLoader
                 Optional.empty(),
                 Optional.empty(),
                 DataSize.of(512, MEGABYTE),
-                false,
                 false,
                 Optional.empty());
