+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHiveConfig.java
@@ -91,9 +91,6 @@ public class TestHiveConfig
                 .setPartitionStatisticsSampleSize(100)
                 .setIgnoreCorruptedStatistics(false)
                 .setCollectColumnStatisticsOnWrite(true)
-                .setS3SelectPushdownEnabled(false)
-                .setS3SelectExperimentalPushdownEnabled(false)
-                .setS3SelectPushdownMaxConnections(500)
                 .setTemporaryStagingDirectoryEnabled(true)
                 .setTemporaryStagingDirectoryPath("/tmp/presto-${USER}")
                 .setDelegateTransactionalManagedTableLocationToMetastore(false)
@@ -177,9 +174,6 @@ public class TestHiveConfig
                 .put("hive.partition-statistics-sample-size", "1234")
                 .put("hive.ignore-corrupted-statistics", "true")
                 .put("hive.collect-column-statistics-on-write", "false")
-                .put("hive.s3select-pushdown.enabled", "true")
-                .put("hive.s3select-pushdown.experimental-textfile-pushdown-enabled", "true")
-                .put("hive.s3select-pushdown.max-connections", "1234")
                 .put("hive.temporary-staging-directory-enabled", "false")
                 .put("hive.temporary-staging-directory-path", "updated")
                 .put("hive.delegate-transactional-managed-table-location-to-metastore", "true")
@@ -260,9 +254,6 @@ public class TestHiveConfig
                 .setPartitionStatisticsSampleSize(1234)
                 .setIgnoreCorruptedStatistics(true)
                 .setCollectColumnStatisticsOnWrite(false)
-                .setS3SelectPushdownEnabled(true)
-                .setS3SelectExperimentalPushdownEnabled(true)
-                .setS3SelectPushdownMaxConnections(1234)
                 .setTemporaryStagingDirectoryEnabled(false)
                 .setTemporaryStagingDirectoryPath("updated")
                 .setDelegateTransactionalManagedTableLocationToMetastore(true)
