+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHiveSplit.java
@@ -79,7 +79,6 @@ public class TestHiveSplit
                         16,
                         ImmutableList.of(createBaseColumn("col", 5, HIVE_LONG, BIGINT, ColumnType.REGULAR, Optional.of("comment"))))),
                 Optional.empty(),
-                false,
                 Optional.of(acidInfo),
                 SplitWeight.fromProportion(2.0)); // some non-standard value
 
@@ -98,7 +97,6 @@ public class TestHiveSplit
         assertEquals(actual.getTableToPartitionMapping().getTableToPartitionColumns(), expected.getTableToPartitionMapping().getTableToPartitionColumns());
         assertEquals(actual.getBucketConversion(), expected.getBucketConversion());
         assertEquals(actual.isForceLocalScheduling(), expected.isForceLocalScheduling());
-        assertEquals(actual.isS3SelectPushdownEnabled(), expected.isS3SelectPushdownEnabled());
         assertEquals(actual.getAcidInfo().get(), expected.getAcidInfo().get());
         assertEquals(actual.getSplitWeight(), expected.getSplitWeight());
     }
