+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestHiveSplitSource.java
@@ -335,7 +335,6 @@ public class TestHiveSplitSource
                     TableToPartitionMapping.empty(),
                     Optional.empty(),
                     Optional.empty(),
-                    false,
                     Optional.empty(),
                     partitionMatchSupplier);
         }
