+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/TestOrcPageSourceMemoryTracking.java
@@ -578,7 +578,6 @@ public class TestOrcPageSourceMemoryTracking
                     TESTING_TYPE_MANAGER,
                     Optional.empty(),
                     Optional.empty(),
-                    false,
                     Optional.empty(),
                     false,
                     NO_ACID_TRANSACTION,
