+++ plugin/trino-hive/src/test/java/io/trino/plugin/hive/orc/TestOrcPredicates.java
@@ -229,7 +229,6 @@ public class TestOrcPredicates
                 TESTING_TYPE_MANAGER,
                 Optional.empty(),
                 Optional.empty(),
-                false,
                 Optional.empty(),
                 false,
                 NO_ACID_TRANSACTION,
